fastapi/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
httpx
python-dotenv
psycopg2-binary
